+++ ProGameCup/core/src/com/mygdx/game/SpaceGame.java
@@ -2,13 +2,11 @@ package com.mygdx.game;
 
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.mygdx.game.screens.MainGameScreen;
 import com.mygdx.game.screens.MainMenuScreen;
-import com.mygdx.game.tiledMapGame.screens.Play;
 
 public class SpaceGame extends Game {
-	public static final int WINDOW_WIDTH = 720;
-	public static final int WINDOW_HEIGHT = 720;
+	public static final int WINDOW_WIDTH = 864;
+	public static final int WINDOW_HEIGHT = 864;
 
 	SpriteBatch batch;
