+++ ProGameCup/core/src/com/mygdx/game/screens/MainGameScreen.java
@@ -5,12 +5,15 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Animation;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
+import com.badlogic.gdx.maps.tiled.TiledMap;
+import com.badlogic.gdx.maps.tiled.TmxMapLoader;
+import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
 import com.mygdx.game.SpaceGame;
-import com.mygdx.game.tiledMapGame.screens.Play;
 
 public class MainGameScreen implements Screen {
 
@@ -20,11 +23,17 @@ public class MainGameScreen implements Screen {
     float x= (float) SpaceGame.WINDOW_HEIGHT /2;
     float y= (float) SpaceGame.WINDOW_WIDTH /2;
     int roll;
+    private float screenWidth, screenHeight;
+    private float mapWidth, mapHeight;
+    private float mapPosX, mapPosY;
+    private TiledMap map;
+    private OrthogonalTiledMapRenderer renderer;
+    private OrthographicCamera camera;
     float stateTime;
     SpriteBatch batch;
     Animation[] rolls;
     public MainGameScreen (SpaceGame game){
-        game.setScreen(new Play(game));
+//        game.setScreen(new Play(game));
         this.game = game;
         batch = game.getBatch();
         walk = new Texture("walk.png");
@@ -39,7 +48,18 @@ public class MainGameScreen implements Screen {
     }
     @Override
     public void show() {
+        TmxMapLoader loader = new TmxMapLoader();
+        map = loader.load("maps/map2.tmx");
+        renderer = new OrthogonalTiledMapRenderer(map);
+        camera = new OrthographicCamera();
+        screenWidth = SpaceGame.WINDOW_WIDTH;
+        screenHeight = SpaceGame.WINDOW_HEIGHT;
 
+        mapWidth = map.getProperties().get("width", Integer.class) * map.getProperties().get("tilewidth", Integer.class);
+        mapHeight = map.getProperties().get("height", Integer.class) * map.getProperties().get("tileheight", Integer.class);
+
+        mapPosX = (screenWidth - mapWidth) / 2;
+        mapPosY = (screenHeight - mapHeight) / 2;
     }
 
     @Override
@@ -63,14 +83,21 @@ public class MainGameScreen implements Screen {
             roll = 2;
         }
         stateTime += delta;
+
+        renderer.setView(camera);
+        renderer.render();
+
         batch.begin();
-        batch.draw((TextureRegion) rolls[roll].getKeyFrame(stateTime, true), x, y, 100, 100);
+        batch.draw((TextureRegion) rolls[roll].getKeyFrame(stateTime, true), x, y, 40, 40);
         batch.end();
     }
 
     @Override
-    public void resize(int i, int i1) {
+    public void resize(int width, int height) {
+        camera.viewportWidth = width;
+        camera.viewportHeight = height;
 
+        camera.update();
     }
 
     @Override
