+++ ProGameCup/desktop/src/com/mygdx/game/DesktopLauncher.java
@@ -2,8 +2,6 @@ package com.mygdx.game;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-import com.mygdx.game.screens.MainGameScreen;
-import com.mygdx.game.tiledMapGame.TiledMapGame;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
